feat(dataset): add GitHub PR–issue dataset builder with quality filters.
Scrapes merged bug-fix PRs into JSONL eval tasks, with mocked API tests and a 15-task built dataset from Textualize/rich.

+++ src/coding_eval/rubric/scorer.py
@@ -27,7 +27,7 @@ class RubricScores:
 
     @property
     def composite(self) -> float:
-        return sum(getattr(self, k) * v for k, v in WEIGHTS.items())
+        return float(sum(getattr(self, k) * v for k, v in WEIGHTS.items()))
 
 
 async def score(
